add WMS 1.3.0 support
