+++ tests/doctests/sos_10_ndbc_getobservation.txt
@@ -28,7 +28,7 @@ GetObservation
     >>> response = ndbc.get_observation(offerings=offerings, responseFormat=responseFormat, observedProperties=observedProperties, eventTime=eventTime)     #doctest: +IGNORE_EXCEPTION_DETAIL
     Traceback (most recent call last):
       File "<stdin>", line 1, in <module>
-    ExceptionReport: 'This is not a valid eventTime!'
+    ServiceException: 'This is not a valid eventTime!'
 
     # NDBC only supports one offering and one observedProperty at a time
 
@@ -51,7 +51,7 @@ DescribeSensor
     >>> response = ndbc.describe_sensor(procedure=procedure, outputFormat=outputFormat)     #doctest: +IGNORE_EXCEPTION_DETAIL
     Traceback (most recent call last):
       File "<stdin>", line 1, in <module>
-    ExceptionReport: 'foobar'
+    ServiceException: 'foobar'
 
     # Valid request
