+++ tests/doctests/wms_getfeatureinfo_130.txt
@@ -1,0 +1,23 @@
+>>> from owslib.wms import WebMapService
+>>> wms = WebMapService('http://geoserv.weichand.de:8080/geoserver/wms', version='1.3.0')
+
+>>> res1 = wms.getfeatureinfo(layers=['bvv:lkr_ex'], srs='EPSG:31468', bbox=(4500000,5500000,4500500,5500500), size=(500,500), format='image/jpeg', info_format="text/html", xy=(250,250))
+>>> html_string1 = res1.read().decode("utf-8")
+>>> ('lkr_ex' in html_string1)
+True
+>>> ('gmd_ex' in html_string1)
+False
+
+>>> res2 = wms.getfeatureinfo(layers=['bvv:lkr_ex','bvv:gmd_ex'], srs='EPSG:31468', bbox=(4500000,5500000,4500500,5500500), size=(500,500), format='image/jpeg', info_format="text/html", xy=(250,250))
+>>> html_string2 = res2.read().decode("utf-8")
+>>> ('lkr_ex' in html_string2)
+True
+>>> ('gmd_ex' in html_string2)
+True
+
+>>> res3 = wms.getfeatureinfo(layers=['bvv:lkr_ex','bvv:gmd_ex'], srs='EPSG:31468', bbox=(4500000,5500000,4500500,5500500), size=(500,500), format='image/jpeg', query_layers=['bvv:lkr_ex'], info_format="text/html", xy=(250,250))
+>>> html_string3 = res3.read().decode("utf-8")
+>>> ('lkr_ex' in html_string3)
+True
+>>> ('gmd_ex' in html_string3)
+False
